products/sigil-generator/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
Pillow==9.0.0
requests==2.26.0
llama-cpp-python==0.2.0 